HAJ-1142 , HAJ-1139, HAJ-1125

+++ shj-admin-portal-services/src/main/java/com/elm/shj/admin/portal/services/dto/CompanyStaffRitualDto.java
@@ -54,7 +54,7 @@ public class CompanyStaffRitualDto implements Serializable {
     @CellIndex(index = 5)
     @SamePerRequest
     @SeasonYear
-    private Long season;
+    private int season;
 
     @CellIndex(index = 6)
     @CompanyCode
